Move UploadUrlPool to a separate file
This was done so that account_info tests can be separated from
upload_url_pool tests, since account_info can't be tested on Jython
due to a lack of sqlite3, but upload_url_pool can be tested with no
issues.
